split RPCMacros into subtraits
